refactor: 컴포넌트명 변경 - ItemListContent → ItemListRenderer
